refactor(Routes): Refactor Routes page
